refactor: config 관련 패키지 생성 및 ErrorMessage 클래스 분리 및 동일 패키지에서만 접근 가능하도록 default 제어자 설정
