+++ src/main/java/devcourse/springbootbasic/config/advice/ErrorMessage.java
@@ -1,0 +1,11 @@
+package devcourse.springbootbasic.config.advice;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+final class ErrorMessage {
+    private final String error;
+    private final String message;
+}
